+++ weblate/trans/autofixes/chars.py
@@ -43,3 +43,19 @@ class RemoveControlChars(AutoFix):
     def fix_single_target(self, target, source, unit):
         result = target.translate(CONTROLCHARS_TRANS)
         return result, result != target
+
+
+class DevanagariDanda(AutoFix):
+    """Fixes Bangla sentence ender."""
+
+    fix_id = "devanadari-danda"
+    name = _("Devanagari danda")
+
+    def fix_single_target(self, target, source, unit):
+        if (
+            unit.translation.language.base_code in ("hi", "bn", "or")
+            and source.endswith(".")
+            and target.endswith((".", "\u09F7", "|"))
+        ):
+            return f"{target[:-1]}\u0964", True
+        return target, False