+++ weblate/trans/tests/test_autofix.py
@@ -9,6 +9,7 @@ from django.test import TestCase
 from weblate.checks.tests.test_checks import MockUnit
 from weblate.trans.autofixes import fix_target
 from weblate.trans.autofixes.chars import (
+    DevanagariDanda,
     RemoveControlChars,
     RemoveZeroSpace,
     ReplaceTrailingDotsWithEllipsis,
@@ -148,3 +149,14 @@ class AutoFixTest(TestCase):
         unit.source = "foo"
         unit.flags = "java-format"
         self.assertEqual(fix.fix_target(["bar'"], unit), (["bar''"], True))
+
+    def test_devanagaridanda(self):
+        non_unit = MockUnit(source="Foo", code="bn")
+        bn_unit = MockUnit(source="Foo.", code="bn")
+        cs_unit = MockUnit(source="Foo.", code="cs")
+        fix = DevanagariDanda()
+        self.assertEqual(fix.fix_target(["Bar."], non_unit), (["Bar."], False))
+        self.assertEqual(fix.fix_target(["Bar."], bn_unit), (["Bar।"], True))
+        self.assertEqual(fix.fix_target(["Bar|"], bn_unit), (["Bar।"], True))
+        self.assertEqual(fix.fix_target(["Bar।"], bn_unit), (["Bar।"], False))
+        self.assertEqual(fix.fix_target(["Bar."], cs_unit), (["Bar."], False))